logging: refactored for using built-in logging
