+++ pyCloudFlareUpdater/__init__.py
@@ -13,4 +13,14 @@
 #
 #     You should have received a copy of the GNU General Public License
 #    along with this program. If not, see <http://www.gnu.org/licenses/>.
-from . import *
+from .values import (
+    LOGGER_NAME,
+    LOG_FILE,
+    LOG_DEFAULT_FORMAT,
+    DEV_CONSOLE_LOG_LEVEL,
+    PRODUCTION_CONSOLE_LOG_LEVEL,
+    DEV_FILE_LOG_LEVEL,
+    PRODUCTION_FILE_LOG_LEVEL
+)
+
+from .logging_utils import init_logging
