+++ pyCloudFlareUpdater/logging_utils/__init__.py
@@ -13,6 +13,4 @@
 #
 #     You should have received a copy of the GNU General Public License
 #    along with this program. If not, see <http://www.gnu.org/licenses/>.
-from ..logging_utils.utils import LoggingHandler
-from ..logging_utils.utils import cleanup_old_logs
-from ..logging_utils.utils import setup_logging
+from .utils import init_logging
